extracted inviation token generation to a concern
