+++ config/application.rb
@@ -15,5 +15,6 @@ module Myflix
       g.orm :active_record
       g.template_engine :haml
     end
+    config.autoload_paths << "#{Rails.root}/lib"
   end
 end
