+++ spec/models/invitation_spec.rb
@@ -5,4 +5,8 @@ describe Invitation do
   it { should validate_presence_of(:recipients_email) }
   it { should validate_presence_of(:recipients_name) }
   it { should validate_presence_of(:message) }
+
+  it_behaves_like 'generates_a_token' do
+    let(:object) { Fabricate(:invitation) }
+  end
 end
