+++ spec/support/shared_examples.rb
@@ -5,3 +5,9 @@ shared_examples 'require_logged_in_user' do
     expect(response).to redirect_to login_path
   end
 end
+
+shared_examples 'generates_a_token' do
+  it 'has a token' do
+    expect(object.invitation_token).to be_present
+  end
+end
